Moves only function from handle_precip.py into NowcastLightning plugin along with associated unit tests.
